move all actix-api cfg tests to test files
